backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-binance==1.0.19
supabase==2.3.0
apscheduler==3.10.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.24.1
sqlalchemy==2.0.23
asyncpg==0.29.0
prometheus-client==0.19.0
sentry-sdk==1.38.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
cryptography==41.0.7
